+++ esphome/components/socket/lwip_raw_tcp_impl.cpp
@@ -11,7 +11,7 @@
 #include "esphome/core/wake.h"
 #include "esphome/core/log.h"
 #ifdef USE_OTA
-#include "esphome/core/ota_wake_hook.h"
+#include "esphome/core/wake.h"  // inline esphome_wake_ota_component_any_context() lives here
 #endif
 
 #ifdef USE_ESP8266
@@ -861,8 +861,8 @@ err_t LWIPRawListenImpl::accept_fn_(struct tcp_pcb *newpcb, err_t err) {
   // Mark the OTA component loop to be re-enabled if it disabled itself while idle.
   // This MUST happen before wake_loop_any_context() below — otherwise the main loop
   // could wake, run a full iteration, and finish before we set the pending-enable
-  // flags, losing the wake event. Inline hook (ota_wake_hook.h) — two volatile stores,
-  // no function call. Safe from RP2040's low-priority user IRQ context.
+  // flags, losing the wake event. Inline hook (wake.h) — two volatile stores, no
+  // function call. Safe from RP2040's low-priority user IRQ context.
   esphome_wake_ota_component_any_context();
 #endif
   // Wake the main loop immediately so it can accept the new connection.
