+++ esphome/core/lwip_fast_select.c
@@ -125,7 +125,7 @@
 #include "esphome/core/lwip_fast_select.h"
 #include "esphome/core/main_task.h"
 #ifdef ESPHOME_USE_OTA
-#include "esphome/core/ota_wake_hook.h"
+#include "esphome/core/wake.h"  // inline esphome_wake_ota_component_any_context() lives here
 #endif
 
 #include <stddef.h>
