+++ esphome/core/wake.h
@@ -3,6 +3,63 @@
 /// @file wake.h
 /// Platform-specific main loop wake primitives.
 /// Always available on all platforms — no opt-in needed.
+///
+/// This file has two sections:
+///   1. A C-compatible section at the top (the inline OTA wake hook) that .c files
+///      like lwip_fast_select.c can include. Uses only <stdbool.h>/<stddef.h>.
+///   2. A C++ section (everything after #ifdef __cplusplus below) with the existing
+///      platform wake primitives and wakeable_delay() implementations.
+
+// ============================================================================
+// C-compatible section: inline OTA wake hook
+// ============================================================================
+//
+// Called from lwip_fast_select.c (and lwip_raw_tcp_impl.cpp / host select) on every
+// NETCONN_EVT_RCVPLUS so a disabled OTA loop can be re-enabled when a monitored
+// socket signals activity. Static inline (not an extern-C trampoline) so the .c
+// callback pays zero function-call overhead — just a pointer load, null check, and
+// two volatile bool stores inlined into the callback body.
+//
+// The two pointers are captured once in Application::set_ota_wake_component(): they
+// point at Component::pending_enable_loop_ and Application::has_pending_enable_loop_requests_.
+// When OTA is not compiled into the build, nothing calls set_ota_wake_component(),
+// the pointers stay NULL, and the inline collapses to a single null check.
+
+#include <stdbool.h>
+#include <stddef.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+// Address of the registered OTA component's pending_enable_loop_ flag, captured at
+// registration time. NULL when no OTA component has registered (including when OTA
+// is not compiled in at all).
+extern volatile bool *esphome_ota_pending_enable_loop_ptr;
+// Address of Application::has_pending_enable_loop_requests_. Set in tandem with the
+// pending_enable pointer above — single null check covers both.
+extern volatile bool *esphome_ota_has_pending_requests_ptr;
+
+// Mark the registered OTA component pending loop-enable. Safe from the LwIP TCP/IP
+// task and raw-TCP IRQ context — only volatile stores, no locks. Callers MUST invoke
+// this BEFORE waking the main task, so the flags are visible on the next iteration.
+static inline void esphome_wake_ota_component_any_context(void) {
+  volatile bool *pending_enable = esphome_ota_pending_enable_loop_ptr;
+  if (pending_enable != NULL) {
+    *pending_enable = true;
+    *esphome_ota_has_pending_requests_ptr = true;
+  }
+}
+
+#ifdef __cplusplus
+}  // extern "C"
+#endif
+
+// ============================================================================
+// C++ section: platform wake primitives
+// ============================================================================
+
+#ifdef __cplusplus
 
 #include "esphome/core/defines.h"
 #include "esphome/core/hal.h"
@@ -124,3 +181,5 @@ inline void wakeable_delay(uint32_t ms) {
 #endif
 
 }  // namespace esphome
+
+#endif  // __cplusplus
